refactor: separate file for processor logic
